+++ CHANGELOG.md
@@ -4,6 +4,44 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2024-12-28
+### ⚠️ Breaking Changes
+- **Updated** the `LoadableViewState.fold` method:
+    - **Renamed** the handler's property in the object parameter from `onError` to `onFailure`. This applies when the `fold` method is called using an object with handler properties.
+
+  ```typescript
+  // Before
+  state.fold({
+      onSuccess: (value) => handleSuccess(value),
+      onError: (error) => handleError(error),
+      onInitial: () => handleInitial(),
+      onLoading: () => handleLoading(),
+  });
+
+  // After
+  state.fold({
+      onSuccess: (value) => handleSuccess(value),
+      onFailure: (error) => handleFailure(error),
+      onInitial: () => handleInitial(),
+      onLoading: () => handleLoading(),
+  });
+  ```
+
+  **Note:** This change does not affect the overload of the `fold` method where handlers are passed as positional arguments.
+
+### 🆕 New Features
+- **Introduced** the `EditableViewState` class to manage mutable states with enhanced state tracking:
+    - Added methods for creating states:
+        - `EditableViewState.initial(value)` for an initial state.
+        - `EditableViewState.loading(current, target)` for a loading state with the current and target values.
+        - `EditableViewState.success(old, succeeded)` for a success state with the previous value and the succeeded value.
+        - `EditableViewState.failure(current, failed, error)` for tracking failure with an associated error and state transition.
+    - Added state-checking methods: `isInitial()`, `isLoading()`, `isSuccess()`, and `isFailure()`.
+    - Added callback methods based on state: `onInitial()`, `onLoading()`, `onSuccess()`, and `onFailure()`.
+    - Implemented `fold()` for transforming the state based on its type.
+
+---
+
 ## [0.0.1] - 2024-11-18
 ### 🚀 Initial Release
 - **Introduced** the `LoadableViewState` class to handle initial, loading, success, and failure states in a view.